requirements.txt
gradio
openai
langchain
pinecone-client
pandas